refactor: PostStorage 관리 로직 SessionStorage 모듈로 이동
